pyproject.toml
[project]
name = "my-agentcore-browser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=0.1.1",
    "playwright>=1.54.0",
    "strands-agents>=1.2.0",
]
